Make ResolveRelative work with all hierarchical URLs.
This makes the ResolveRelative method work correctly for all
hierarchical URLs while preserving the special support for the
"standard" schemes.

The change is needed for Android WebView, where it is perfectly
valid for an embedder to load a data URL with a base URL that is
a custom scheme (like myapp://resources/) and then later intercept
those requests.

Review=https://codereview.chromium.org/11367010/
BUG=http://crbug.com/159832
Patch by Martin Kosiba



git-svn-id: http://google-url.googlecode.com/svn/trunk@181 8873c55e-713a-0410-88f8-23d9c3d90b1b